Implement light tracking using world time
This adds support for using different time and calendar modules for
tracking time in the campaign, and integrating that with torches.
Allows for example to use SimpleCalendar module to move time forward
when the party is spending extra time searching a room.

By default behavior is unchanged, but with a new option real time
tracking can be disabled in cases where GM wants to use clock from
another module.

Also changed the inventory and light tracker use same rounding for the
time remaining.