added nominal ability tabbing
This new tabbing modifies the old behavior from left-to-right and top-to-bottom tabbing with top-to-bottom then left-to-right without changing the appearance of the npc and player sheets.

This is to follow the normal order ability scores are listed in ShadowDark. This is especially useful for when GM's need to manual enter many monsters statistics which are listed in official
 publications as S, D, C, I, W, Ch.